requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
argostranslate==1.4.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
click==7.1.2
csnake==0.3.4
ctranslate2==1.20.1
DAWG-Python==0.7.2
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.18
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.0.0
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
neo4j==4.3.1
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.21.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py-translate==1.0.3
pyarrow==4.0.1
pycparser==2.20
pydeck==0.6.2
Pygments==2.9.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
rfc3986==1.5.0
Send2Trash==1.7.1
sentencepiece==0.1.94
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
stanza==1.1.1
streamlit==0.84.1
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
torch==1.9.0
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
